refactor(plugin): add file to consolidate chrome/firefox APIs
At 2-3 places in the codebase, we're checking if the browser
is chrome or firefox, and hen using the corresponding APIs
chrome.* or browser.* respectively. To reduce code repetition
this adds a file browser-compat.js that exposes objects corresponding
to various browser APIs such that they can be used by other functions
in a browser agnostic way.